Restructure plugin public facing pages
This commit includes the addition of extra pages, one to cover privacy
and a further more indepth cookie page. There was also a refactor of
some of the file naming and methods which we introduce breaking changes
that will have to either mitgate against or uninstall and reinstall the
plugin to activate.