MVP variant B cookie banner
This is a draft version of the variant B cookie banner, following Toni's design. There is keyboard focus trapping on the expanding banner, it's been tested with VoiceOver, and has been tested with Browser Stack in Edge, IE, Safari and mobile devices.

There are a couple of visual bugs - the toggle buttons don't completely mirror the designs, and there is an issue with the background continuing to scroll when the banner is up. I've figured these things are minor bugs that whilst might be annoying to a few users, they won't prevent anyone from being able to use it, and so they're okay for MVP. I will add a note of them to the Trello card for us to fix on whichever banner we choose after running this experiment.