Fix save button and increase z-index on variant b cookie banner
The save button the cookie banner did set analytics cookies, but it didn't set the cookie to tell it to stop showing the banner. This fixes this bug. There is also an issue with some items with a high z-index on the site appearing on top of the cookie banner, so this increases the z-index to prevent this.