Remove use of rems, and adds govuk class to close button
Rems are currently causing issues on PPJ, which we need to put the banner on for a WCAG audit. This turns them off, so the banner works on PPJ. This won't negatively affect other sites using the banner.

It also adds a class to the close button, to fix a styling issue (without the class, it wasn't appearing big enough).