Abstract out gov-uk classes and small bug fixes
This fixes visual bugs in IE and with the table captions. It also removes govuk classes from the HTML and adds to them to the SASS instead. This makes it less likely to pick up styling from the theme, and helps encapsulate it a bit more. The ccfw classes are much less likely to be in another theme than the govuk- classname.