Cap max modal content width at 1100px
The DAC report mentioned a usability issue to do with the distance of the toggle from the text. It looks like the screenshot is referring to a user who's using ZoomText and has a very big screen, and on a very big screen, the toggle floats out to the right-hand side of the modal. This PR puts a limit on it, so it'll always be near the text no matter the screen size.