Remove dependency on GDS Transport font
The cookie banner uses the Design System, which out of the box wants to use the Transport font. This requires a license though, which we don't have, and are unlikely to be procuring for the banner. So, this adds in a SASS variable to stop it looking for it, and to stop the console errors that were happening as a result.