Move focus to modal, not close button, on modal appearance
When the modal pops up, the focus previously went to the close button, but this triggers the focus state on the close button and looks a bit strange. A user might also end up closing it by accident. Instead,  I've moved it to the banner itself, so that the first focusable element is the close button, but that you're not already on it.