Fix leaky 100% on HTML and body
When the cookie banner pops up, the body and HTML elements get set to 100%, so that you can't scroll the background. This is for accessibility reasons. However, this was previously being applied with CSS, and still applying when the banner wasn't open. This hasn't caused issues on other sites yet, but on IMA it was affecting how the background colour was appearing. So, this tidies it up - that class is now toggled with JS, so it only applies when the banner is open.