Adjust spacings on cookie modal
Toni and I went through and looked at the spacings on the modal in the cookie banner. This adds in the changes that came out of that discussion.

It also adds in another breakpoint - the table that expand out from the details sections were breaking on mobile, as there wasn't enough space. This adjusts the paddings so that there is enough room for the words, and the smaller breakpoint means that that affects fewer screen sizes - it looks a bit crampt.