Fix broken cookie toggle
The cookie toggle was not working properly for screen reader users. The aria-labelledby wasn't working, so it wasn't reading out what you were toggling. It also wasn't saying anything other than 'off', because it was using aria-pressed instead of aria-checked. This corrects both of those issues.

It also fixes a visual bug in Safari, where the font color wasn't turning white (which meant that the button always said off on it).