Add JS to toggle banner with button
This PR adds in the JS. When you click the cookie icon button, the banner appears, and the cookie icon button disappears. When you then close the banner, the cookie icon button reappears.

Still to do - update cookie choices, when user makes a different choice, on second time interacting with the banner.