Fix bug where cookies weren't deleting (and corresponding visual issue)
Somewhere in the previous commits, the JS that removed the cookies got altered. This meant that the banner wasn't deleting GA cookies when requested. This has now been fixed. This PR also fixes an issue where the toggle state CSS wasn't always matching its state, that resulted from this change.