Simplify banner - bring all on one page
I'd previously started to build this so that the options were spread across multiple pages. After thinking about it a bit more though, I figured that really they can all fit on one page.

So this gets rid of the multiple tabs, and brings the GA checkbox on the main page. It also adds in toggles for Twitter, Vimeo and YouTube - these are still a work in progress.

Issues still to do:
- Get checkboxes to show as checked if they have been checked, on settings page
- Add in copy for them to toggle (TBC)